Drops Strapdown/Strapdown-Zeta
Strapdown is convenient but has some performance drawbacks that made
rendering the pages quite frankly, ugly.

This is a overhaul of the html frontend, that dropped Strapdown and
replaced it with Showdownjs (a simple markdown2html converter) and
some javascript to generate the html.

Bootstrap is still imported and magically makes the page prettier. We
do lose the ability to pick a theme, but this could be reintroduced
later if needed.

The html templates have been reviewed to include proper head links, a
navbar, and an inheritance scheme.

Finally, every reference to Strapdown has been removed from the docs
and the build scripts.