Make page load more elegant
Because of the showdown modifies the rendering of the page, there's a
split second on load where the unstyled markdown was shown to the user
before being blogstrapped.

This commit hides this behavior by turning off the visibility of the
body until all the rendering is done.

It also avoids messing up the order of js loading before this.