Separates ArticleReader into content and metadata
Metadata are still, as of this commit, empty. I still need to have a
proper parser going through the articles and extracting metadata. It
should come in future commits.

In the meantime this commit does some minor hygenic changes to prepare
for the future work. Most notably, `render_markdown` now renders a
(mostly empty) Flask template.