Refactors handling of configuration context
* Provides a default value to all the mandatory config variables

This object also serves as documentation on what are the values that a
user can override in their conf

* Provides a helper function to unify how context is passed

For now we're passing the same values in the context of all the
templates we render. We lose a bit of flexibility, but the code is
prettier and easier to handle.