Adds the Metadata Tags feature
Metadata tags are any line at the top of the article in the pattern
of:
	`# <key>: <value>

More information on the legal format of the metadata can be found in
the submitted tests.

Metadata are removed from the article before serving it. The metadata
parser effectively splits any article into metadata/content. Only the
content part is served by Blogstrap.

Undocumented, we also get the ability to override any of the (silent
so far) Page Variables.